Update Citus tests for Citus 11.1.
In Citus 11.1 we need to make sure the nodes metadata is in sync before some
cluster wide operations such as DROP TABLE. This is best done by calling a
citus test only function: wait_until_metadata_sync.

The function needs to first be created/exposed at the SQL level.

+++ tests/test_nonha_citus_operation.py
@@ -17,7 +17,9 @@ def setup_module():
 
 
 def teardown_module():
-    coordinator1b.run_sql_query("DROP TABLE t1")
+    if coordinator1b is not None:
+        coordinator1b.run_sql_query("select public.wait_until_metadata_sync()")
+        coordinator1b.run_sql_query("DROP TABLE t1")
     cluster.destroy()
 
 
@@ -46,6 +48,10 @@ def test_002_init_coordinator():
 
     print()  # make the debug output more readable
     assert coordinator1a.wait_until_state(target_state="single")
+
+    # we need to expose some Citus testing internals
+    assert coordinator1a.wait_until_pg_is_running()
+    coordinator1a.create_wait_until_metadata_sync()
 
 
 def test_003_init_workers():
